Revert "Revert "EES-1039 - separated out Admin-only functions out of shared FootnoteService into Admin-specific FootnoteService""
This reverts commit c98c599f48b4b1b14ce7ffede8e2edc906ad1947.
