+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/ManageContent/FootnoteService.cs
@@ -1,0 +1,335 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using GovUk.Education.ExploreEducationStatistics.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Data.Model.Database;
+using GovUk.Education.ExploreEducationStatistics.Data.Model.Services;
+using GovUk.Education.ExploreEducationStatistics.Data.Model.Services.Interfaces;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using IFootnoteService = GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.ManageContent.IFootnoteService;
+
+namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.ManageContent
+{
+    public class FootnoteService : AbstractRepository<Footnote, Guid>, IFootnoteService
+    {
+        private readonly IFilterService _filterService;
+        private readonly IFilterGroupService _filterGroupService;
+        private readonly IFilterItemService _filterItemService;
+        private readonly IIndicatorService _indicatorService;
+        private readonly ISubjectService _subjectService;
+
+        public FootnoteService(StatisticsDbContext context,
+            ILogger<FootnoteService> logger,
+            IFilterService filterService,
+            IFilterGroupService filterGroupService,
+            IFilterItemService filterItemService,
+            IIndicatorService indicatorService,
+            ISubjectService subjectService) : base(context, logger)
+        {
+            _filterService = filterService;
+            _filterGroupService = filterGroupService;
+            _filterItemService = filterItemService;
+            _indicatorService = indicatorService;
+            _subjectService = subjectService;
+        }
+
+        public Footnote CreateFootnote(string content,
+            IEnumerable<Guid> filterIds,
+            IEnumerable<Guid> filterGroupIds,
+            IEnumerable<Guid> filterItemIds,
+            IEnumerable<Guid> indicatorIds,
+            IEnumerable<Guid> subjectIds)
+        {
+            var footnote = DbSet().Add(new Footnote
+            {
+                Content = content,
+                Filters = new List<FilterFootnote>(),
+                FilterGroups = new List<FilterGroupFootnote>(),
+                FilterItems = new List<FilterItemFootnote>(),
+                Indicators = new List<IndicatorFootnote>(),
+                Subjects = new List<SubjectFootnote>()
+            }).Entity;
+
+            CreateSubjectLinks(footnote, subjectIds);
+            CreateFilterLinks(footnote, filterIds);
+            CreateFilterGroupLinks(footnote, filterGroupIds);
+            CreateFilterItemLinks(footnote, filterItemIds);
+            CreateIndicatorsLinks(footnote, indicatorIds);
+
+            _context.SaveChanges();
+            return footnote;
+        }
+
+        public void DeleteFootnote(Guid id)
+        {
+            var footnote = Find(id, new List<Expression<Func<Footnote, object>>>
+            {
+                f => f.Filters, f => f.FilterGroups, f => f.FilterItems, f => f.Indicators, f => f.Subjects
+            });
+
+            DeleteEntities(footnote.Subjects);
+            DeleteEntities(footnote.Filters);
+            DeleteEntities(footnote.FilterGroups);
+            DeleteEntities(footnote.FilterItems);
+            DeleteEntities(footnote.Indicators);
+
+            Remove(id);
+            _context.SaveChanges();
+        }
+
+        public Footnote GetFootnote(Guid id)
+        {
+            return DbSet().Where(footnote => footnote.Id == id)
+                .Include(footnote => footnote.Filters)
+                .ThenInclude(filterFootnote => filterFootnote.Filter)
+                .Include(footnote => footnote.FilterGroups)
+                .ThenInclude(filterGroupFootnote => filterGroupFootnote.FilterGroup)
+                .ThenInclude(filterGroup => filterGroup.Filter)
+                .Include(footnote => footnote.FilterItems)
+                .ThenInclude(filterItemFootnote => filterItemFootnote.FilterItem)
+                .ThenInclude(filterItem => filterItem.FilterGroup)
+                .ThenInclude(filterGroup => filterGroup.Filter)
+                .Include(footnote => footnote.Indicators)
+                .ThenInclude(indicatorFootnote => indicatorFootnote.Indicator)
+                .ThenInclude(indicator => indicator.IndicatorGroup)
+                .Include(footnote => footnote.Subjects)
+                .SingleOrDefault();
+        }
+
+        public Footnote UpdateFootnote(Guid id,
+            string content,
+            IEnumerable<Guid> filterIds,
+            IEnumerable<Guid> filterGroupIds,
+            IEnumerable<Guid> filterItemIds,
+            IEnumerable<Guid> indicatorIds,
+            IEnumerable<Guid> subjectIds)
+        {
+            var footnote = Find(id, new List<Expression<Func<Footnote, object>>>
+            {
+                f => f.Filters,
+                f => f.FilterGroups,
+                f => f.FilterItems,
+                f => f.Indicators,
+                f => f.Subjects
+            });
+
+            DbSet().Update(footnote);
+
+            footnote.Content = content;
+
+            UpdateFilterLinks(footnote, filterIds);
+            UpdateFilterGroupLinks(footnote, filterGroupIds);
+            UpdateFilterItemLinks(footnote, filterItemIds);
+            UpdateIndicatorLinks(footnote, indicatorIds);
+            UpdateSubjectLinks(footnote, subjectIds);
+
+            _context.SaveChanges();
+            return GetFootnote(id);
+        }
+
+        public IEnumerable<Footnote> GetFootnotes(Guid releaseId)
+        {
+            return DbSet().Where(footnote =>
+                    (!footnote.Subjects.Any() ||
+                     footnote.Subjects.Any(subjectFootnote => subjectFootnote.Subject.ReleaseId == releaseId))
+                    && (!footnote.Filters.Any() || footnote.Filters.Any(filterFootnote =>
+                            filterFootnote.Filter.Subject.ReleaseId == releaseId))
+                    && (!footnote.FilterGroups.Any() || footnote.FilterGroups.Any(filterGroupFootnote =>
+                            filterGroupFootnote.FilterGroup.Filter.Subject.ReleaseId == releaseId))
+                    && (!footnote.FilterItems.Any() || footnote.FilterItems.Any(filterItemFootnote =>
+                            filterItemFootnote.FilterItem.FilterGroup.Filter.Subject.ReleaseId == releaseId))
+                    && (!footnote.Indicators.Any() || footnote.Indicators.Any(indicatorFootnote =>
+                            indicatorFootnote.Indicator.IndicatorGroup.Subject.ReleaseId == releaseId)))
+                .Include(footnote => footnote.Filters)
+                .ThenInclude(filterFootnote => filterFootnote.Filter)
+                .Include(footnote => footnote.FilterGroups)
+                .ThenInclude(filterGroupFootnote => filterGroupFootnote.FilterGroup)
+                .ThenInclude(filterGroup => filterGroup.Filter)
+                .Include(footnote => footnote.FilterItems)
+                .ThenInclude(filterItemFootnote => filterItemFootnote.FilterItem)
+                .ThenInclude(filterItem => filterItem.FilterGroup)
+                .ThenInclude(filterGroup => filterGroup.Filter)
+                .Include(footnote => footnote.Indicators)
+                .ThenInclude(indicatorFootnote => indicatorFootnote.Indicator)
+                .ThenInclude(indicator => indicator.IndicatorGroup)
+                .Include(footnote => footnote.Subjects);
+        }
+
+        private void CreateSubjectLinks(Footnote footnote, IEnumerable<Guid> subjectIds)
+        {
+            var subjects = _subjectService.FindMany(subject => subjectIds.Contains(subject.Id));
+
+            var links = footnote.Subjects;
+            foreach (var subject in subjects)
+            {
+                AddSubjectLink(ref links, footnote.Id, subject.Id);
+            }
+        }
+
+        private void CreateFilterLinks(Footnote footnote, IEnumerable<Guid> filterIds)
+        {
+            var filters = _filterService.FindMany(filter => filterIds.Contains(filter.Id));
+
+            var links = footnote.Filters;
+            foreach (var filter in filters)
+            {
+                AddFilterLink(ref links, footnote.Id, filter.Id);
+            }
+        }
+
+        private void CreateFilterGroupLinks(Footnote footnote, IEnumerable<Guid> filterGroupIds)
+        {
+            var filterGroups = _filterGroupService.FindMany(filterGroup => filterGroupIds.Contains(filterGroup.Id),
+                new List<Expression<Func<FilterGroup, object>>>
+                {
+                    filterGroup => filterGroup.Filter
+                });
+
+            var links = footnote.FilterGroups;
+            foreach (var filterGroup in filterGroups)
+            {
+                AddFilterGroupLink(ref links, footnote.Id, filterGroup.Id);
+            }
+        }
+
+        private void CreateFilterItemLinks(Footnote footnote, IEnumerable<Guid> filterItemIds)
+        {
+            var filterItems = _filterItemService.FindMany(filterItem => filterItemIds.Contains(filterItem.Id),
+                new List<Expression<Func<FilterItem, object>>>
+                {
+                    filterItem => filterItem.FilterGroup.Filter
+                });
+
+            var links = footnote.FilterItems;
+            foreach (var filterItem in filterItems)
+            {
+                AddFilterItemLink(ref links, footnote.Id, filterItem.Id);
+            }
+        }
+
+        private void CreateIndicatorsLinks(Footnote footnote, IEnumerable<Guid> indicatorIds)
+        {
+            var indicators = _indicatorService.FindMany(indicator => indicatorIds.Contains(indicator.Id),
+                new List<Expression<Func<Indicator, object>>>
+                {
+                    indicator => indicator.IndicatorGroup
+                });
+
+            var links = footnote.Indicators;
+            foreach (var indicator in indicators)
+            {
+                AddIndicatorLink(ref links, footnote.Id, indicator.Id);
+            }
+        }
+
+        private void UpdateFilterLinks(Footnote footnote, IEnumerable<Guid> filterIds)
+        {
+            if (!SequencesAreEqualIgnoringOrder(
+                footnote.Filters.Select(link => link.FilterId), filterIds))
+            {
+                footnote.Filters = new List<FilterFootnote>();
+                CreateFilterLinks(footnote, filterIds);
+            }
+        }
+
+        private void UpdateFilterGroupLinks(Footnote footnote, IEnumerable<Guid> filterGroupIds)
+        {
+            if (!SequencesAreEqualIgnoringOrder(
+                footnote.FilterGroups.Select(link => link.FilterGroupId), filterGroupIds))
+            {
+                footnote.FilterGroups = new List<FilterGroupFootnote>();
+                CreateFilterGroupLinks(footnote, filterGroupIds);
+            }
+        }
+
+        private void UpdateFilterItemLinks(Footnote footnote, IEnumerable<Guid> filterItemIds)
+        {
+            if (!SequencesAreEqualIgnoringOrder(
+                footnote.FilterItems.Select(link => link.FilterItemId), filterItemIds))
+            {
+                footnote.FilterItems = new List<FilterItemFootnote>();
+                CreateFilterItemLinks(footnote, filterItemIds);
+            }
+        }
+
+        private void UpdateIndicatorLinks(Footnote footnote, IEnumerable<Guid> indicatorIds)
+        {
+            if (!SequencesAreEqualIgnoringOrder(
+                footnote.Indicators.Select(link => link.IndicatorId), indicatorIds))
+            {
+                footnote.Indicators = new List<IndicatorFootnote>();
+                CreateIndicatorsLinks(footnote, indicatorIds);
+            }
+        }
+
+        private void UpdateSubjectLinks(Footnote footnote, IEnumerable<Guid> subjectIds)
+        {
+            if (!SequencesAreEqualIgnoringOrder(
+                footnote.Subjects.Select(link => link.SubjectId), subjectIds))
+            {
+                footnote.Subjects = new List<SubjectFootnote>();
+                CreateSubjectLinks(footnote, subjectIds);
+            }
+        }
+
+        private static void AddIndicatorLink(ref ICollection<IndicatorFootnote> links, Guid footnoteId, Guid linkId)
+        {
+            links.Add(new IndicatorFootnote
+            {
+                FootnoteId = footnoteId,
+                IndicatorId = linkId
+            });
+        }
+
+        private static void AddFilterLink(ref ICollection<FilterFootnote> links, Guid footnoteId, Guid linkId)
+        {
+            links.Add(new FilterFootnote
+            {
+                FootnoteId = footnoteId,
+                FilterId = linkId
+            });
+        }
+
+        private static void AddFilterGroupLink(ref ICollection<FilterGroupFootnote> links, Guid footnoteId, Guid linkId)
+        {
+            links.Add(new FilterGroupFootnote
+            {
+                FootnoteId = footnoteId,
+                FilterGroupId = linkId
+            });
+        }
+
+        private static void AddFilterItemLink(ref ICollection<FilterItemFootnote> links, Guid footnoteId, Guid linkId)
+        {
+            links.Add(new FilterItemFootnote
+            {
+                FootnoteId = footnoteId,
+                FilterItemId = linkId
+            });
+        }
+
+        private static void AddSubjectLink(ref ICollection<SubjectFootnote> links, Guid footnoteId, Guid subjectId)
+        {
+            links.Add(new SubjectFootnote
+            {
+                FootnoteId = footnoteId,
+                SubjectId = subjectId
+            });
+        }
+
+        private void DeleteEntities<T>(IEnumerable<T> entitiesToDelete)
+        {
+            foreach (var t in entitiesToDelete)
+            {
+                _context.Entry(t).State = EntityState.Deleted;
+            }
+        }
+
+        private static bool SequencesAreEqualIgnoringOrder(IEnumerable<Guid> left, IEnumerable<Guid> right)
+        {
+            return left.OrderBy(id => id).SequenceEqual(right.OrderBy(id => id));
+        }
+    }
+}
